+++ R/regon.R
@@ -61,7 +61,7 @@ StatRegon <- ggproto('StatRegon', Stat,
       p * 2 * pi
     }))
     data$group <- paste0(data$group, '_', seq_len(nrow(data)))
-    data <- data[rep(data$group, data$sides), ]
+    data <- data[rep(seq_len(nrow(data)), data$sides), ]
     x_tmp <- sin(pos) * data$r
     y_tmp <- cos(pos) * data$r
     data$x <- data$x0 + x_tmp * cos(data$angle) - y_tmp * sin(data$angle)